AG: fixed save dirs.

+++ Dungeon of hell/Menus/SingleplayerInGameMenu/SingleplayerInGameMenuViewModel.cs
@@ -31,11 +31,20 @@ namespace Dungeon_of_hell
             MainMenuView = new RelayCommand(() =>
             {
                 Audio_player.Play("menuSelect");
-                Directory.Delete(GlobalSettings.Settings.AssetsPath + "save\\"+"Props",true);
+                if (Directory.Exists(GlobalSettings.Settings.AssetsPath + "save\\" + "Props"))
+                {
+                    Directory.Delete(GlobalSettings.Settings.AssetsPath + "save\\" + "Props", true);
+                }
+                if (Directory.Exists(GlobalSettings.Settings.AssetsPath + "save\\" + "Items"))
+                {
+                    Directory.Delete(GlobalSettings.Settings.AssetsPath + "save\\" + "Items", true);
+                }
+                if (Directory.Exists(GlobalSettings.Settings.AssetsPath + "save\\" + "Enemys"))
+                {
+                    Directory.Delete(GlobalSettings.Settings.AssetsPath + "save\\" + "Enemys", true);
+                }
                 Directory.CreateDirectory(GlobalSettings.Settings.AssetsPath + "save\\" + "Props");
-                Directory.Delete(GlobalSettings.Settings.AssetsPath + "save\\" + "Items", true);
                 Directory.CreateDirectory(GlobalSettings.Settings.AssetsPath + "save\\" + "Items");
-                Directory.Delete(GlobalSettings.Settings.AssetsPath + "save\\" + "Enemys", true);
                 Directory.CreateDirectory(GlobalSettings.Settings.AssetsPath + "save\\" + "Enemys");
                 foreach (Item item in GetViewProperty<List<Item>>("Singleplayer", "Items"))
                 {
